+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 - Fix Watch-only connection and login
 - Fix reloading transactions and subaccounts
+- Fix transaction blockheight 
 
 ## [3.8.7] - 2022-09-06